Cleanup NTP SI/SR generation step
Removed code duplication and made them as util method.
