+++ lib/ntpUtil.js
@@ -4,7 +4,34 @@
 
 const childProcess = require('child_process')
 const fs = require('fs')
+const path = require('path')
 const util = require('../lib/util')
+const request = require('request')
+
+const jsonSchemaVersion = 1
+
+const createPhotoJsonFile = (path, body) => {
+  fs.writeFileSync(path, body)
+}
+
+const getImageFileNameListFrom = (photoJsonObj) => {
+  let fileList = []
+  if (photoJsonObj.logo)
+    fileList.push(photoJsonObj.logo.imageUrl)
+
+  if (photoJsonObj.wallpapers) {
+    photoJsonObj.wallpapers.forEach((wallpaper) => {
+      fileList.push(wallpaper.imageUrl)
+    })
+  }
+
+  if (photoJsonObj.topSites) {
+    photoJsonObj.topSites.forEach((topSiteObj) => {
+      fileList.push(topSiteObj.iconUrl)
+    })
+  }
+  return fileList
+}
 
 const generatePublicKeyAndID = (privateKeyFile) => {
   childProcess.execSync(`openssl rsa -in ${privateKeyFile} -pubout -out public.pub`)
@@ -29,6 +56,59 @@ const generatePublicKeyAndID = (privateKeyFile) => {
   }
 }
 
+const prepareAssets = (jsonFileUrl, targetResourceDir, targetJsonFileName) => {
+  return new Promise(function (resolve, reject) {
+    let jsonFileBody = '{}'
+
+    // Download and parse jsonFileUrl.
+    // If it doesn't exist, create with empty object.
+    request(jsonFileUrl, async function (error, response, body) {
+      if (error) {
+        console.error(`Error from ${jsonFileUrl}:`, error)
+        return reject(error)
+      }
+      if (response && response.statusCode === 200) {
+        jsonFileBody = body
+      }
+      let photoData = {}
+      try {
+        photoData = JSON.parse(jsonFileBody)
+      } catch (err) {
+        console.error(`Invalid json file ${jsonFileUrl}`)
+        return reject(error)
+      }
+      // Make sure the data has a schema version so that clients can opt to parse or not
+      const incomingSchemaVersion = photoData.schemaVersion
+      if (!incomingSchemaVersion) {
+        console.error(`Error: Cannot parse JSON data at ${jsonFileUrl} since it doesn't has a schema version. This region will not be updated.`)
+        return reject(error)
+      } else if (incomingSchemaVersion !== jsonSchemaVersion) {
+        // We don't support this file format
+        console.error(`Error: Cannot parse JSON data at ${jsonFileUrl} since it has a schema version of ${incomingSchemaVersion} but we expected ${jsonSchemaVersion}! This region will not be updated.`)
+        return reject(error)
+      }
+
+      createPhotoJsonFile(path.join(targetResourceDir, 'photo.json'), JSON.stringify(photoData))
+
+      // Download image files that specified in jsonFileUrl
+      const imageFileNameList = getImageFileNameListFrom(photoData)
+      const downloadOps = imageFileNameList.map((imageFileName) => new Promise(resolve => {
+        const targetImageFilePath = path.join(targetResourceDir, imageFileName)
+        const targetImageFileUrl = new URL(imageFileName, jsonFileUrl).href
+        request(targetImageFileUrl)
+          .pipe(fs.createWriteStream(targetImageFilePath))
+          .on('finish', () => {
+            console.log(targetImageFileUrl)
+            resolve()
+          })
+      }))
+      await Promise.all(downloadOps)
+      resolve()
+    })
+  })
+}
+
 module.exports = {
-  generatePublicKeyAndID
+  generatePublicKeyAndID,
+  prepareAssets
 }
