+++ scripts/generateNTPSponsoredImages.js
@@ -5,12 +5,9 @@
 const path = require('path')
 const mkdirp = require('mkdirp')
 const fs = require('fs-extra')
-const request = require('request')
 const commander = require('commander')
 const util = require('../lib/util')
-
-const jsonFileName = 'photo.json'
-const jsonSchemaVersion = 1
+const ntpUtil = require('../lib/ntpUtil')
 
 const getRegionList = () => {
   return [ 'AF', 'AL', 'DZ', 'AS', 'AD', 'AO', 'AI', 'AQ', 'AG', 'AR', 'AM', 'AW', 'AU', 'AT', 'AZ', 'BS',
@@ -29,78 +26,6 @@ const getRegionList = () => {
            'SJ', 'SZ', 'SE', 'CH', 'TW', 'TJ', 'TZ', 'TH', 'TL', 'TG', 'TK', 'TO', 'TT', 'TN', 'TR', 'TM',
            'TC', 'TV', 'UG', 'UA', 'AE', 'GB', 'US', 'UM', 'UY', 'UZ', 'VU', 'VE', 'VN', 'VG', 'VI', 'WF',
            'EH', 'YE', 'ZM', 'ZW' ]
-}
-
-const createPhotoJsonFile = (path, body) => {
-  fs.writeFileSync(path, body)
-}
-
-const getImageFileNameListFrom = (photoJsonObj) => {
-  let fileList = []
-  if (photoJsonObj.logo)
-    fileList.push(photoJsonObj.logo.imageUrl)
-
-  if (photoJsonObj.wallpapers) {
-    photoJsonObj.wallpapers.forEach((wallpaper) => {
-      fileList.push(wallpaper.imageUrl)
-    })
-  }
-  return fileList
-}
-
-function downloadForRegion (jsonFileUrl, targetResourceDir) {
-  return new Promise(function (resolve, reject) {
-    const jsonFilePath = path.join(targetResourceDir, jsonFileName)
-    let jsonFileBody = '{}'
-
-    // Download and parse photo.json.
-    // If it doesn't exist, create with empty object.
-    request(jsonFileUrl, async function (error, response, body) {
-      if (error) {
-        console.error(`Error from ${jsonFileUrl}:`, error)
-        return reject(error)
-      }
-      if (response && response.statusCode === 200) {
-        jsonFileBody = body
-      }
-      let photoData = {}
-      try {
-        photoData = JSON.parse(jsonFileBody)
-      } catch (err) {
-        console.error(`Invalid json file ${jsonFileUrl}`)
-        return reject(error)
-      }
-      // Make sure the data has a schema version so that clients can opt to parse or not
-      const incomingSchemaVersion = photoData.schemaVersion
-      if (!incomingSchemaVersion) {
-        // Source has no schema version, assume and set current version.
-        // TODO(petemill): Don't allow this once the source is established to always
-        // have a schema version.
-        photoData.schemaVersion = jsonSchemaVersion
-      } else if (incomingSchemaVersion !== jsonSchemaVersion) {
-        // We don't support this file format
-        console.error(`Error: Cannot parse JSON data at ${jsonFileUrl} since it has a schema version of ${incomingSchemaVersion} but we expected ${jsonSchemaVersion}! This region will not be updated.`)
-        return reject(error)
-      }
-
-      createPhotoJsonFile(jsonFilePath, JSON.stringify(photoData))
-
-      // Download image files that specified in photo.json
-      const imageFileNameList = getImageFileNameListFrom(photoData)
-      const downloadOps = imageFileNameList.map((imageFileName) => new Promise(resolve => {
-        const targetImageFilePath = path.join(targetResourceDir, imageFileName)
-        const targetImageFileUrl = new URL(imageFileName, jsonFileUrl).href
-        request(targetImageFileUrl)
-          .pipe(fs.createWriteStream(targetImageFilePath))
-          .on('finish', () => {
-            console.log(targetImageFileUrl)
-            resolve()
-          })
-      }))
-      await Promise.all(downloadOps)
-      resolve()
-    })
-  })
 }
 
 function getTargetRegionList(targetRegions, excludedTargetRegions) {
@@ -124,8 +49,8 @@ async function generateNTPSponsoredImages (dataUrl, targetRegions, excludedTarge
     console.log(`Downloading ${region}...`)
     const targetResourceDir = path.join(rootResourceDir, region)
     mkdirp.sync(targetResourceDir)
-    const jsonFileUrl = `${dataUrl}${region}/${jsonFileName}`
-    await downloadForRegion(jsonFileUrl, targetResourceDir)
+    const sourceJsonFileUrl = `${dataUrl}${region}/photo.json`
+    await ntpUtil.prepareAssets(sourceJsonFileUrl, targetResourceDir)
   }
 }
